test(ReservationTimeService): 예약 가능 시간 생성 검증 테스트
